+++ sobyaninSergey/components/screens/lab1.js
@@ -1,0 +1,36 @@
+import { useState } from "react";
+import { TouchableOpacity, View, Dimensions } from "react-native";
+
+const Lab1 = () => {
+  const [size, setSize] = useState(100);
+  const windowWidth = Dimensions.get("window").width;
+  const windowHeight = Dimensions.get("window").height;
+
+  return (
+    <View
+      style={{
+        flex: 1,
+        justifyContent: "center",
+        alignItems: "center",
+        backgroundColor: "black",
+      }}
+    >
+      <TouchableOpacity
+        onPress={() => {
+          if (size > windowWidth || size > windowHeight) {
+            return setSize(100);
+          }
+          return setSize(size + 50);
+        }}
+        style={{
+          width: size,
+          height: size,
+          borderRadius: size / 2,
+          backgroundColor: "#004524",
+        }}
+      />
+    </View>
+  );
+};
+
+export default Lab1;
